pyproject.toml
[project]
name = "trenni"
version = "0.1.0"
description = "Minimal supervisor for Palimpsest job orchestration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "click>=8.0",
    "pyyaml>=6.0",
    "uuid-utils>=0.14.1",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "yoitsu-contracts",
]

[project.scripts]
trenni = "trenni.cli:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.uv]
package = true

[tool.uv.sources]
yoitsu-contracts = { path = "../yoitsu-contracts" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
